chore: refactor mother cache to interface + implementation
